% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classilist.R
\name{classilist}
\alias{classilist}
\title{Export Data if valid and Launch Classilist application}
\usage{
classilist(dataSet, location, features, predicted, actual, probs)
}
\arguments{
\item{dataSet}{classified data having features, class probabilities, Target class and Predicted class for each entry}

\item{location}{- path of classilist Installation (This should contain data directory for classilist to access data)}

\item{features}{Vector of feature names}

\item{predicted}{String Column name of predicted class}

\item{actual}{String Column name of Target class}

\item{probs}{Vector of Class Probability column names}
}
\value{
modified dataSet if successful, else the original dataSet
}
\description{
Export Data if valid and Launch Classilist application
}
\examples{
classilist(data , "/home/user/abc.csv" , c("column1" , "column2" , ..), "predicted_column",
     "actual_column" , c("class1" , "class2" , ... ))
}

