% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateClasses.R
\name{validateClasses}
\alias{validateClasses}
\title{Matches Probability Columns with Classes}
\usage{
validateClasses(classes, probs)
}
\arguments{
\item{classes}{from Target Column of data set}

\item{probs}{Vector of Class Probability column names}
}
\value{
True : if all classes have probabilities
False : otherwise
}
\description{
Matches Probability Columns with Classes
}
\examples{
validateClasses( c("class1" , "class2" , ..), c("probClass1" , "probClass2"))
}

