% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateColumns.R
\name{validateColumns}
\alias{validateColumns}
\title{Validates all column categories}
\usage{
validateColumns(coln, features, predicted, actual, probs)
}
\arguments{
\item{features}{Vector of feature names}

\item{predicted}{String Column name of predicted class}

\item{actual}{String Column name of Target class}

\item{probs}{Vector of Class Probability column names}
}
\value{
True : if the column name format is correct
False : otherwise
}
\description{
Validates all column categories
}
\examples{
validateColumns( c("column1" , "column2" , ..), "predicted_column","actual_column" , c("class1" , "class2" , ... ))
}

