% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateFile.R
\name{validateFile}
\alias{validateFile}
\title{Validates File name and path}
\usage{
validateFile(filename)
}
\arguments{
\item{fileName}{String filename for validation}
}
\value{
False : if path is invalid and/or doesn't seems to be correct classilist Installation
True : if path exists and is correct Installation Directory for classilist
}
\description{
Validates File name and path
}
\examples{
validateFile("/home/user/classilist/")
}

